Create a menu page object in cypress
