+++ cypress/integration/formats.spec.js
@@ -1,5 +1,5 @@
 import { teardown, logoutAndLoginAs } from '../support/authentication';
-import * as homePage from '../support/homePage';
+import * as menu from '../support/menu';
 import * as datasetImportPage from '../support/datasetImportPage';
 import * as configureField from '../support/configureField';
 import * as searchDrawer from '../support/searchDrawer';
@@ -9,8 +9,8 @@ describe('Transformers & Formats', () => {
 
     describe('LIST format', () => {
         it('should display a composed field with a LIST format', () => {
-            homePage.openAdvancedDrawer();
-            homePage.goToAdminDashboard();
+            menu.openAdvancedDrawer();
+            menu.goToAdminDashboard();
 
             datasetImportPage.importDataset('dataset/simple.csv');
             datasetImportPage.setUriColumnValue();
@@ -23,7 +23,7 @@ describe('Transformers & Formats', () => {
             datasetImportPage.publish();
 
             datasetImportPage.goToPublishedResources();
-            searchDrawer.openSearchDrawer();
+            menu.openSearchDrawer();
 
             cy.contains('Row 1').should('be.visible');
             cy.contains('Row 2').should('be.visible');
@@ -32,8 +32,8 @@ describe('Transformers & Formats', () => {
 
     describe('Broken Formats & Wrong Values', () => {
         beforeEach(() => {
-            homePage.openAdvancedDrawer();
-            homePage.goToAdminDashboard();
+            menu.openAdvancedDrawer();
+            menu.goToAdminDashboard();
 
             datasetImportPage.importDataset('dataset/broken-formats.csv');
             datasetImportPage.importModel('model/broken-formats.json');
