+++ cypress/integration/graph_page.spec.js
@@ -1,14 +1,14 @@
 import { teardown } from '../support/authentication';
 import * as datasetImportPage from '../support/datasetImportPage';
-import * as homePage from '../support/homePage';
+import * as menu from '../support/menu';
 import * as graphPage from '../support/graphPage';
 import * as searchDrawer from '../support/searchDrawer';
 
 describe('Graph Page', () => {
     beforeEach(() => {
         teardown();
-        homePage.openAdvancedDrawer();
-        homePage.goToAdminDashboard();
+        menu.openAdvancedDrawer();
+        menu.goToAdminDashboard();
 
         datasetImportPage.importDataset('dataset/chart.csv');
         datasetImportPage.importModel('model/chart.json');
@@ -17,9 +17,9 @@ describe('Graph Page', () => {
     });
 
     it('should not reset filters & search query when switching to another graph', () => {
-        homePage.openChartDrawer();
+        menu.openChartDrawer();
 
-        homePage.goToChart('Bar Chart');
+        menu.goToChart('Bar Chart');
         graphPage.getStats().should('have.text', 'Found 50 on 50');
 
         graphPage.searchFor('Biodiversity');
@@ -28,8 +28,8 @@ describe('Graph Page', () => {
         graphPage.setFacet('Publication Year', '2011');
         graphPage.getStats().should('have.text', 'Found 4 on 50');
 
-        homePage.openAdvancedDrawer();
-        homePage.goToChart('Bubble Chart');
+        menu.openAdvancedDrawer();
+        menu.goToChart('Bubble Chart');
 
         graphPage.getSearchInput().should('have.value', 'Biodiversity');
         graphPage.getStats().should('have.text', 'Found 4 on 50');
@@ -42,16 +42,16 @@ describe('Graph Page', () => {
     });
 
     it('should copy filters to search drawer', () => {
-        homePage.openChartDrawer();
+        menu.openChartDrawer();
 
-        homePage.goToChart('Bar Chart');
+        menu.goToChart('Bar Chart');
 
         graphPage.searchFor('Biodiversity');
         graphPage.getStats().should('have.text', 'Found 5 on 50');
         graphPage.setFacet('Publication Year', '2011');
         graphPage.getStats().should('have.text', 'Found 4 on 50');
 
-        searchDrawer.openSearchDrawer();
+        menu.openSearchDrawer();
         searchDrawer.getFacet('Publication Year').click();
         searchDrawer
             .getFacetItem('Publication Year', '2011')
@@ -59,7 +59,7 @@ describe('Graph Page', () => {
             .should('not.checked');
         searchDrawer.searchInput().should('have.value', '');
 
-        searchDrawer.closeSearchDrawer();
+        menu.closeSearchDrawer();
         graphPage.browseResults();
 
         searchDrawer.getFacet('Publication Year').click();
